commit-1 28/02/24

+++ Qomoi/src/main/java/com/qomoi/dto/CourseLocationResponse.java
@@ -34,6 +34,6 @@ public class CourseLocationResponse {
 
     private List<TrainerResponse> trainer;
 
-    private float courseAmt;
+    private double courseAmt;
 
 }
